Changed Task1

+++ homework3/src/Task1.java
@@ -3,10 +3,10 @@ public class Task1 {
         int x = 47;
         int y = 33;
         int z = 10;
-        System.out.println((double) getAverage(x, y, z));
+        System.out.println(getAverage(x, y, z));
     }
 
-    static int getAverage(int x, int y, int z){
-        return x > z ? (x + y) / 2 : z;
+    static double getAverage(int x, int y, int z){
+        return x > z ? (double) (x + y) / 2 : z;
     }
 }
